preprocess: factored out EumssiConverter
